+++ src/Pages/Dict/Dict.input.ts
@@ -1,13 +1,37 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsNumber, IsOptional, IsString } from 'class-validator';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+import { IsArray, IsNumber, IsOptional, IsString } from 'class-validator';
 
 export class DictInput {
   @IsNumber()
   @ApiProperty({
     description: 'ID словаря',
-    example: 1,
+    example: 18,
   })
   id: number;
+
+  @IsOptional()
+  @IsString()
+  @ApiPropertyOptional({
+    description: 'Название статуса',
+    example: 'Отмена приказа',
+  })
+  name?: string;
+
+  @IsOptional()
+  @IsArray()
+  @ApiPropertyOptional({
+    description: 'ID словарей, которые не нужно включать в результат',
+    example: [],
+  })
+  not_in_ids?: number[];
+
+  @IsOptional()
+  @IsArray()
+  @ApiPropertyOptional({
+    description: 'Названия словарей, которые не нужно включать в результат',
+    example: [],
+  })
+  not_in_names?: string[];
 }
 
 export class GetRawDictNamesClass {