+++ src/Pages/Dict/Dict.service.ts
@@ -2,7 +2,7 @@ import { Dict, DictName } from '@contact/models';
 import { InjectModel } from '@sql-tools/nestjs-sequelize';
 import { Injectable } from '@nestjs/common';
 import { DictInput, GetRawDictNamesClass } from './Dict.input';
-import { Op } from 'sequelize';
+import { Op } from '@sql-tools/sequelize';
 @Injectable()
 export class DictService {
   constructor(
@@ -12,14 +12,23 @@ export class DictService {
   ) {}
   async dict(body: DictInput) {
     return await this.ModelDict.findAll({
-      where: { parent_id: body.id },
+      attributes: ['id', 'code', 'name', 'r_code', 'typ'],
+      where: {
+        parent_id: body.id,
+        ...(body.name && {
+          [Op.and]: [{ name: { [Op.like]: `%${body.name}%` } }],
+        }),
+        ...(body.not_in_ids && { code: { [Op.notIn]: body.not_in_ids } }),
+        ...(body.not_in_names && {
+          name: { [Op.notIn]: body.not_in_names },
+        }),
+      },
     });
   }
 
   async getRawDictNames({ name, code }: GetRawDictNamesClass) {
     const dictNames = await this.modelDictName.findAll({
       attributes: ['id', 'code', 'name', 'r_code', 'typ'],
-      //@ts-expect-error expecting error
       where: {
         code: code ? code : undefined,
         name: name ? { [Op.like]: `%${name}%` } : undefined,