+++ src/Pages/RejectStatuses/RejectStatuses.service.ts
@@ -8,6 +8,11 @@ import { DebtRejectStatuses } from 'src/Modules/Database/local.database/models/D
 import { InjectModel } from '@sql-tools/nestjs-sequelize';
 import LawActRejectStatuses from 'src/Modules/Database/local.database/models/LawActRejectStatuses.model';
 import { Dict } from '@contact/models';
+import { Op } from '@sql-tools/sequelize';
+import CacheService from 'src/Modules/Cache/Cache.service';
+import { CacheEnums } from 'src/utils/enums';
+import { GetAllResponse } from './interfaces/get-all-response';
+import { endPerfomance } from 'src/utils/loggers';
 
 @Injectable()
 export class RejectStatusesService {
@@ -18,29 +23,76 @@ export class RejectStatusesService {
     private readonly modelDebtRejectStatuses: typeof DebtRejectStatuses,
     @InjectModel(LawActRejectStatuses, 'local')
     private readonly modelLawActRejectStatuses: typeof LawActRejectStatuses,
-
     @InjectModel(Dict, 'contact')
     private readonly modelDict: typeof Dict,
+    private readonly cacheService: CacheService,
   ) {}
 
-  async getAll(): Promise<{
-    debtRejectStatuses: number[];
-    lawActRejectStatuses: string[];
-  }> {
+  async resetCacheRejectStatuses() {
+    const newRejectStatuses = await this.getAllFromDB();
+    this.cacheService.set({
+      key: CacheEnums.REJECT_STATUSES,
+      value: newRejectStatuses,
+      ttl: 60000,
+    });
+  }
+
+  async getAll() {
+    const begin = performance.now();
+    const cachedData = await this.cacheService.get({
+      key: CacheEnums.REJECT_STATUSES,
+    });
+    if (cachedData) {
+      endPerfomance(begin, 'From Cache'.cyan);
+      return cachedData as GetAllResponse;
+    }
+    this.logger.log('Cache not found, getting from database');
+    const response = await this.getAllFromDB();
+    return response;
+  }
+
+  async getAllFromDB(): Promise<GetAllResponse> {
     try {
       const debtRejectStatuses = await this.modelDebtRejectStatuses.findAll();
       const lawActRejectStatuses =
         await this.modelLawActRejectStatuses.findAll();
 
-      const response: {
-        debtRejectStatuses: number[];
-        lawActRejectStatuses: string[];
-      } = {
-        debtRejectStatuses: debtRejectStatuses.map((debt) => debt.reject_id),
-        lawActRejectStatuses: lawActRejectStatuses.map(
+      const attributes = ['id', 'code', 'name'];
+
+      const dict_for_debt = await this.modelDict.findAll({
+        where: {
+          parent_id: 6,
+          code: {
+            [Op.in]: [...debtRejectStatuses.map((debt) => debt.reject_id)],
+          },
+        },
+        attributes,
+      });
+
+      const dict_for_law_act = await this.modelDict.findAll({
+        where: {
+          parent_id: 18,
+          name: {
+            [Op.in]: [...lawActRejectStatuses.map((law) => law.reject_name)],
+          },
+        },
+        attributes,
+      });
+
+      const response: GetAllResponse = {
+        debt_reject_statuses: debtRejectStatuses.map((debt) => debt.reject_id),
+        dict_for_debt,
+        law_act_reject_statuses: lawActRejectStatuses.map(
           (law) => law.reject_name,
         ),
+        dict_for_law_act,
       };
+
+      this.cacheService.set({
+        key: CacheEnums.REJECT_STATUSES,
+        value: response as GetAllResponse,
+        ttl: 60000,
+      });
 
       return response;
     } catch (error) {
@@ -50,20 +102,32 @@ export class RejectStatusesService {
   }
 
   async addDebtRejectStatus(reject_id: number): Promise<DebtRejectStatuses> {
+    const DEBT_DICT_CODE = 6; // Код словаря для долга
     try {
       const dict = await this.modelDict.findOne({
         where: {
-          code: reject_id,
+          [Op.and]: [{ parent_id: DEBT_DICT_CODE }, { code: reject_id }],
         },
       });
       if (!dict) {
         throw new BadRequestException('Статус отказа не найден');
       }
-      const debtRejectStatus = await this.modelDebtRejectStatuses.create({
-        reject_id: dict.id,
+      const element = await this.modelDebtRejectStatuses.findOne({
+        where: {
+          reject_id,
+        },
+      });
+      if (element) {
+        throw new BadRequestException(
+          `Статус отказа debt - "${dict.name}" уже существует`,
+        );
+      }
+      const debt_reject_status = await this.modelDebtRejectStatuses.create({
+        reject_id: dict.code,
       });
 
-      return debtRejectStatus;
+      await this.resetCacheRejectStatuses();
+      return debt_reject_status;
     } catch (error) {
       this.logger.error(error);
       const err_message =
@@ -78,12 +142,34 @@ export class RejectStatusesService {
   async addLawActRejectStatus(
     reject_name: string,
   ): Promise<LawActRejectStatuses> {
+    const LAW_ACT_DICT_CODE = 18; // Код словаря для законаS
     try {
-      const lawActRejectStatus = await this.modelLawActRejectStatuses.create({
-        reject_name,
+      const dict = await this.modelDict.findOne({
+        where: {
+          [Op.and]: [{ parent_id: LAW_ACT_DICT_CODE }, { name: reject_name }],
+        },
       });
+      if (!dict) {
+        throw new BadRequestException('Статус отказа law_act не найден');
+      }
+      const element = await this.modelLawActRejectStatuses.findOne({
+        where: {
+          reject_name,
+        },
+      });
+      if (element) {
+        throw new BadRequestException(
+          `Статус отказа law_act - "${dict.name}" уже существует`,
+        );
+      }
+      const law_act_reject_status = await this.modelLawActRejectStatuses.create(
+        {
+          reject_name,
+        },
+      );
 
-      return lawActRejectStatus;
+      await this.resetCacheRejectStatuses();
+      return law_act_reject_status;
     } catch (error) {
       this.logger.error(error);
       const err_message =
@@ -93,5 +179,29 @@ export class RejectStatusesService {
         err_message,
       );
     }
+  }
+
+  async deleteDebtRejectStatus(reject_id: number): Promise<DebtRejectStatuses> {
+    const debt_reject_status = await this.modelDebtRejectStatuses.findOne({
+      where: { reject_id },
+    });
+    if (!debt_reject_status) {
+      throw new BadRequestException('Статус отказа не найден');
+    }
+    await debt_reject_status.destroy();
+    await this.resetCacheRejectStatuses();
+    return debt_reject_status;
+  }
+
+  async deleteLawActRejectStatus(name: string): Promise<LawActRejectStatuses> {
+    const law_act_reject_status = await this.modelLawActRejectStatuses.findOne({
+      where: { reject_name: name },
+    });
+    if (!law_act_reject_status) {
+      throw new BadRequestException('Статус отказа не найден');
+    }
+    await law_act_reject_status.destroy();
+    await this.resetCacheRejectStatuses();
+    return law_act_reject_status;
   }
 }