+++ src/Pages/RejectStatuses/dto/debt-reject-status.dto.ts
@@ -2,13 +2,12 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
 import { IsNotEmpty, IsNumber } from 'class-validator';
 
-export class AddDebtRejectStatusDto {
+export class DebtRejectStatusDto {
   @ApiProperty({
     description: 'The ID of the reject status',
     example: 1,
   })
   @IsNumber({}, { message: 'reject_id must be a number' })
   @IsNotEmpty({ message: 'reject_id is required' })
-  @Transform(({ value }) => Number(value))
   reject_id: number;
 }