+++ src/main.ts
@@ -1,4 +1,4 @@
-import { ValidationPipe } from '@nestjs/common';
+import { Logger, ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import {
   FastifyAdapter,
@@ -7,8 +7,7 @@ import {
 import { AppModule } from './app.module';
 import https from './utils/https';
 import client from './utils/client';
-import { contentParser } from 'fastify-multer';
-import { LocalSeed } from './Modules/Database/local.database/local.seed';
+// import { contentParser } from 'fastify-multer';
 import moment from 'moment';
 import './utils/CustomCA';
 import 'colors';
@@ -17,6 +16,7 @@ import {
   getSwaggerOptions,
   getSwaggerOptionsCustom,
 } from './utils/getSwaggerOptions';
+import multipart from '@fastify/multipart';
 
 export const node = process.env.NODE_ENV;
 class bootstrapOptions {
@@ -33,18 +33,24 @@ class bootstrapOptions {
 
 moment.tz.setDefault('GMT');
 async function bootstrap() {
+  const logger = new Logger('bootstrap');
   const options = new bootstrapOptions();
   const app = await NestFactory.create<NestFastifyApplication>(
     AppModule,
     options.adapter,
   );
-  app.register(contentParser);
-  // try {
-  //   await app.get(LocalSeed).seed();
-  // } catch (e) {
-  //   console.log(e);
-  //   throw e;
-  // }
+
+  await app
+    .register(multipart, {
+      limits: {
+        fileSize: 10 * 1024 * 1024,
+      },
+    })
+    .then(() => {
+      logger.debug('multipart registered');
+    });
+
+  // app.register(contentParser);
   const config = new DocumentBuilder()
     .setTitle('Подача')
     .setDescription('ПО Подача для отдела ИП')
@@ -62,7 +68,23 @@ async function bootstrap() {
   );
   SwaggerModule.setup('docs', app, document, getSwaggerOptionsCustom());
   app.useGlobalPipes(new ValidationPipe({ transform: true }));
-  app.enableCors();
+
+  const cors_config =
+    node === 'dev'
+      ? {
+          origin: true,
+          methods: ['GET', 'POST', 'PUT', 'DELETE', 'PATCH', 'OPTIONS'],
+          allowedHeaders: [
+            'Content-Type',
+            'Authorization',
+            'token',
+            'Content-Length',
+          ],
+          credentials: true,
+        }
+      : {};
+  console.log('cors_config'.yellow, cors_config);
+  app.enableCors(cors_config);
   await app.listen(client('port'), '0.0.0.0');
   console.log(
     `NODE_ENV: ${node}, Send-Application is running on: ${await app.getUrl()}/docs