+++ src/utils/enums.ts
@@ -1,0 +1,3 @@
+export enum CacheEnums {
+  REJECT_STATUSES = 'reject_statuses',
+}